requirements.txt
# Core multi-agent framework
autogen-agentchat>=0.4.0
autogen-ext[magentic-one]>=0.4.0

# Azure AI / Foundry
azure-ai-projects>=1.0.0b5
azure-identity>=1.19.0
openai>=1.55.0

# Web UI
streamlit>=1.40.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
